Loading files for challenge 8
Hi, I'm loading the existing files for the challenge 8 requirements.  I've not been able to complete the database creation, which I will likely loop back to if I'm able to find some time this weekend.  Submitting what I have for now to gauge the current score.  I will likely continue to work on the assignments to improve complete the challenge if time allows.  Thanks so much!!